Cooldown Changes
Cooldowns fixed

+++ wServer/realm/StatsManager.cs
@@ -72,7 +72,7 @@ public class StatsManager
     {
         if (Owner.HasConditionEffect(ConditionEffects.Dazed))
             return MinAttackFreq;
-        var rof = MinAttackFreq + (this[5] / 75.0f) * (MaxAttackFreq - MinAttackFreq);
+        var rof = MinAttackFreq + this[5] / (float)75.0f * (MaxAttackFreq - MinAttackFreq);
         if (Owner.HasConditionEffect(ConditionEffects.Berserk))
             rof *= 1.5f;
         return rof;
